Cannot delete company if user exist check added

+++ app/routers/company.py
@@ -7,7 +7,7 @@ from app.crud import company as crud_company
 from app.schemas import company as schemas
 from app.database import get_db  # your DB session function under database
 from app.crud.company import get_all_company, get_company_by_id, create_company, delete_company, crud_get_users_by_company
-from app.models import Company
+from app.models import Company,User
 
 router = APIRouter(
     prefix="/company",
@@ -45,6 +45,10 @@ def deletecompany(company_id: int,db: Session = Depends(get_db)):
     company_exist = db.query(Company).filter(Company.id == company_id).first()
     if not company_exist:
         raise HTTPException(status_code=404, detail="Company doesn't exist")
+
+    user_exist = db.query(User).all()
+    if user_exist:
+        raise HTTPException(status_code=400, detail="User already exists")
 
     delete_company(db, company_id)
     return {"Company Deleted Successfully"}
